Standardize Lambda Handler Name
Why these changes are being introduced:
After some conversation, we have decided on a standard naming pattern
for Python Lambda@Edge functions in our CDN.

How this addresses that need:
* Add ADR#1 to capture the correct naming pattern for the main Python
file and the function that is executed
* Replace all instances of `lambda_handler` with just `handler` in the
code and tests
* Update README

Side effects of this change:
None. Note that the code currenly in this repository does nothing! The
actual Python code that is running in Dev1 and Stage-Workloads is a
copy of the manually created jp-cdn-auth function from Dev1.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/GDT-106